requirements.txt
# Adding peep so we can track it on requires.io. For packages
# installation we use peep.py shipped in ./bin
# sha256: KoBM4H9Zz1WtVFuy4WMSwRNkuU0_k4bW4SFFsuOOXBw
peep==2.4.1

# sha256: c1CauJkYk_5byimUihn9wtv5xN1KDpiqnFgDvveKsiY
# sha256: zpqSSG7aW_PZ5Ah_gCSSsVtxsGuJTWgM-Wb1wnlD5TI
# sha256: m7S7kSEJqteivOjiKGp9mgnrzKrv5Br4VYZrZgcVkGA
peepin==0.11

# sha256: ZExcwtBko41DlidUnmOCqLwouAokP-XLukCx79K6ves
# sha256: LRdOSj9UcI0NW2_xv1SucWUug7sG11drOyDZFrKcNlM
Django==1.8.5

# sha256: 8uJz7TSsu1YJYtXPEpF5NtjfAil98JvTCJuFRtRYQTg
# sha256: ygF2j97N4TQwHzFwdDIm9g7f9cOTXxJDc3jr2RFQY1M
dj-database-url==0.3.0

# sha256: as-au751fvddwuzZ2RunSVR5Qauv--af8ghqnjfUkEw
psycopg2==2.6.1

# sha256: ufv3R_qOcR0zDy9Da5suz17tnrZ6Y3yoGrzRHCq-3sg
python-decouple==2.3

# sha256: LiSsXQBNtXFJdqBKwOgMbfbkfpjDVMssDYL4h51Pj9s
Jinja2==2.7.3

# sha256: pOwa_1m5WhS0XrLiN2GgF56YMZ2lp-t2tW6ozce4ccM
MarkupSafe==0.23

# sha256: pRec3pItK04EXuWxHocyPud9NjrkApT8glLyXWoOrwY
# sha256: i8g1CCiCrZoBLNeQxGABHk2WvzUS2YoE09q75FOToIk
gunicorn==19.3.0

# sha256: VYVza7U0XcnLcEwSkq997IUSZmqw-RbG8mduDwYCMeg
# sha256: kpfCKE84KWGnvdmJeFlixWKX1HYUJT57vdglG2nUrOQ
whitenoise==2.0.2

# sha256: 0yV2uq13_dnmhZvpfCtqPBem9IQ2Q4nbcx8Gl7ehlmE
django-csp==2.0.3

# sha256: mWbSw_hrIzdNqYkFhPFrA_bkYFWCTjChCHFpdNRdHkU
# sha256: 98W-TbhvGnbMxMgH-QCFNidavsPigV0YJl1pSu7zyzw
# sha256: bOkY5I2u1knhOqTrLOmdgCEkEUAb5CKIyURzqBEJJE4
# sha256: ZEHnmPiGN0TJBXTeJTvtMoxHf9LIbK9SuNnYXx7Ln5Y
# sha256: TWTtG54OcwlfXPqH8Ol920yEAEno7-t-Y7RhGLodYjo
# sha256: Oe8FQeSJsc9VFc7fGx-Z3lhf1mBGPNzeQ6A9fZIEJ4I
# sha256: aFsZlAJlchd0WK8UPMJKIIL0APGOY_3I4u1o3ZO7D0k
# sha256: hEYBPfvemz9Pm_6glBjd1SZ2CoeYxBut0_cHqYrQdFs
# sha256: 1f9wYInpFQdg0ji2YgOiTUoKYXUWBBAZGZX_BNGJP6o
# sha256: p4tITVRy3Yxoj4s-7hhkaiXGbORbLCZlKFD2r5zlKxc
# sha256: xO5wy0B_koRResNo8SHPB5anE0uWHlPZ2vGqro9E-5A
# sha256: APNfsOfi_DSM0JHSXTbGdotyFGPRI5bWipSRqKX5Hs8
pytz==2015.4

# sha256: WbUVu327zdMH-BesOIVOgIOev6M8XbeX9lRQeTyUXA0
newrelic==2.52.0.40

# sha256: j8ToZaKfPG7K5AGn3OGCR_ir0EbSTjkOFq-Zdsyceqg
# sha256: Ln675Z2MheYm422Z8NsvV4OUMT0_fOncnx2lfvbNdTc
flake8==2.4.1

# sha256: SqEp342QB7GSv4IBP0FVM5lGUtfKqTDQAmh-tCpsKkE
# sha256: uLfjVjC1U54moZffxgBb6eHpoTVJazd3I6jrwBuby_8
pep8==1.6.2

# sha256: Bd9YSinu6poqIRDdNi5T0E4MS7F1S01xI09lGRfzwvA
# sha256: AmkcI85pnyUodLfCfxTPJuPU6CtY5dWE8AC3q1vjal8
pyflakes==0.9.2

# sha256: vWwID7NyrrywzhnjXdrHRPKr9ae--iB9stEJfUjv5jo
# sha256: X36m-zqpr-FG0H_W1c7feIdH2LDCnkRzJFPCstsePRY
mccabe==0.3.1

# sha256: VkWzEfLrzStU3NmTwLatLrXK2Q9fL-PUXBhMhD9E1Ac
# sha256: 5nkMaUKPpsPg5NYz4mrReJcORouv_gOphikFS9uEMI4
# sha256: jmS0QeFimMCAJbgmEmste8XBd20tLwcWchZvYV8yeIc
# sha256: MfXU5lrA5yp-r8Mey5WrfJgbtJ7TFLaTEbykXGq4-ss
ipython==3.2.0

# sha256: F2PRVkET9euJ33eHmo0yEyc8TX_5Pcs3owcM3ww0_Xw
ipdb==0.8.1

# sha256: Ld8Y2jsGIfpD_uS3KQ2grnibRvuJkniorMzaGVxJeac
# sha256: Gm5RMMK0LS3jAWk8KZ94zEvTUB54thDAjkXvxw4rURQ
Sphinx==1.3.1

# sha256: _5spJogkCW_JETkEYF2D9qqBcQLA8XARW0s58K4GUfM
# sha256: dPYz7Tph2h0dWcMYVIPIGp1zRr8Oe18prQdkpvFZtoo
sphinx_rtd_theme==0.1.8

# sha256: Egr4uJMGHscr5KqRuTRV5tDUqAF1wHpTgZkQxNiXSDA
# sha256: CMxlW-wlusPmYKstHFFHHUUy7zk_xA-vvSUBQCY1Frs
sphinx-autobuild==0.5.2

# sha256: 3OvUkoESYxYm9MTQ31l4fHSEBOZt2pUhEAMOqIPTuM0
# sha256: x9txeBCraWX2bIzwOYqYydjfmC2jm0zX8WKRHriVlvo
docutils==0.12

# sha256: CjoiZeHvsN76ci0fQpcw3J35da3DxgrH0rQyvfA8BB0
# sha256: Xe0vqQlP19_rPakmNkCf1wKg0H1gYoNQTX7gRAHO5cs
# sha256: cyCRkITm2sj0VAY4pGRHo71zD8oXKvwX0sA-7SLPT1E
Pygments==2.0.2

# sha256: b0wujFraFwOf8cHji8Ub6pMXrzZ24oRmFIyN9fx48TI
django-jinja==1.4.1

# sha256: 67kjSmrxku6ahiO8lapstSNVdj0E_u2xmoiM2kC7NGY
django-allauth==0.23.0

# sha256: ksUcPs7IRsvsSu_xH5_0cwPUpj-TsOasDsAqCR_tcO8
# sha256: wtEz5H4KdwXJJy7vANegnBdPW_F6En_tjixkmVVsx4I
python-openid==2.2.5

# sha256: 70v-RmPKO5eplYYMAXO5Z-vZgDPQLzjJ4bLLtsGR2a0
oauthlib==1.0.3

# sha256: IPl2zc4CpCtpzoDp4DiXpRgUs21EizcohUYIbrxHMUY
# sha256: OYo9ttYYmdJf1KBsbKEgUbDOFx1wXezX7VURUXtLuT0
requests==2.7.0

# sha256: rVC-9UD-VRikZTw4IEUqiBtqBCyw-Lt2V8SRxr02VLs
Pillow==3.0.0
# sha256: R_NHKksF6A24fEqYrFz2XLd8owE3uXB98SCzb3XUIWA
opbeat==3.1.3
# sha256: fRdUe2UhbMXG-8BK7lUIjM1ZF8B3UwTZb3AXwmx4nNc
# sha256: AMxHk1r7vYMmD90oOwqnkOZY0qcZIgSfbkZ9yooSRTc
django-filter==0.11.0
